### 2024-06-14 #201 @matteobachetti
Stingray Submission
Submitting Author: Name @matteobachetti
All current maintainers: @dhuppenkothen @mgullik @jdswinbank @matteolucchini1
Package Name: Stingray
One-Line Description of Package: A spectral-timing software package for astrophysical X-ray (and other) data
Repository Link:   https://github.com/stingraysoftware/stingray

Version submitted: 2.1
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11383212.svg)](https://doi.org/10.5281/zenodo.11383212)
JOSS DOI: 10.21105/joss.07389
Version accepted: 2.1

---

Stingray is a Python library for "spectral timing", i.e. time-series analysis techniques that can be used to study how variability changes or correlates between different energy bands/wavelengths. It is Astropy-affiliated, and with an ever growing user base now comprising hundreds of researchers around the globe.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

This package is mostly focused on the analysis of new data from high-energy missions. We added data validation and testing because it can be used as a quick look tool to point out possible anomalies in observations (e.g. solar or other background flares).

The target audience is principally researchers and students of X-ray and multi-wavelength astronomy.
This package fills a void for a free and open source (spectral-)timing package for X-ray astronomy. XRONOS, formerly maintained by HEASARC, is currently unmaintained, and the analysis of high-energy timeseries is done mostly with proprietary software or mission-specific packages. We provide a Python package that eases the learning curve for newcomers, also thanks to extensive tutorials based on Jupyter notebooks, and provides experts with a powerful, robust library for their analysis. We provide software to analyze astronomical time series and do a number of things, including periodicity searches, time lag calculations, covariance spectra, power spectral modeling.

Our package is arguably the most well-known Python package for X-ray spectral timing.

https://github.com/pyOpenSci/software-submission/issues/195

- [ x contains a README with instructions for installing the development version.

- [ ] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/`.  **We have a previous JOSS paper describing the package. Is it acceptable to have a new paper, including the many improvements made in the last five years?**
