### 2024-06-19 fdf01c5
Specify lightcurve chunks
