### 2024-07-01 #201 @cmarmo
Hi there! I'm Chiara and I am taking care of the Editor in chief role for the next three months.
Thank you for submitting your package for pyOpenSci review.
Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

---
---

Stingray is in excellent form! 💪
I am going to check the availability of our Astropy editors and we will be back to you shortly.
