### 2024-07-02 #201 @hamogu
## Editor response to review:

---

## Editor comments
:wave: Hi @reviewer-one and @reviewer-two! Thank you for volunteering to review
for pyOpenSci! <Add any additional banter here that you wish..>

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: <Insert deadline DATE HERE>
---

Reviewers:
Due date:
