### 2024-07-02 #201 @hamogu
>  We have a previous JOSS paper describing the package. Is it acceptable to have a new paper, including the many improvements made in the last five years?

It's up to JOSS, not us, how exactly they handle that. Yo save everyone time, JOSS will generally accept our review as theirs, because there is really no need for a second set of reviewers that do essentially the same checks that our reviewers just did. However the final decision to accept into JOSS is up to JOSS.

So, I think there are two ways this can be done, at out choice:

-  Either you contact JOSS and ask for their opinion. (In that case, I would pause this review until you hear back from them.)
- Or you write the `paper.md` as you would like to have it, and we review the code and paper as we normally do and pass it to JOSS once our review is done. (In that case, I would pause this review until you've written the `paper.md` so that our reviewers can actually review the paper.)

Option 2 is faster (you don't have to wait for the JOSS response first), but carries the risk that the JOSS editor decides not to publish it later.

Note that JOSS says: "we also review submissions that are significant contributions made to existing packages".
Based on this, I suggest option 2 (you add a 'paper.md` here before we review) and I will specifically instruct our reviewers to comment on if they judge the development in the last 5 years to be a "significant contribution".
