### 2024-07-10 #201 @matteobachetti
@hamogu the paper is ready, but I'm still missing a couple OrcIDs and affiliations. Does it have to be merged or the link to the PR is sufficient?
