+++ stingray/pulse/tests/test_accelsearch.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pytest
 from stingray.pulse.accelsearch import accelsearch
-from stingray.utils import HAS_NUMBA
 
 pytestmark = pytest.mark.slow
 
