+++ stingray/tests/test_lombscargle.py
@@ -5,7 +5,6 @@ import pytest
 from scipy.interpolate import interp1d
 
 from stingray.events import EventList
-from stingray.exceptions import StingrayError
 from stingray.lightcurve import Lightcurve
 from stingray.lombscargle import LombScargleCrossspectrum, LombScarglePowerspectrum
 from stingray.lombscargle import _autofrequency
