### 2024-07-19 a50a73b
add get approximate psd
