### 2024-07-19 27da9f1
add diagnostics for the approximation
