+++ stingray/modeling/gpmodeling.py
@@ -3,6 +3,9 @@ import matplotlib.pyplot as plt
 import functools
 from stingray import Lightcurve
 
+from matplotlib.lines import Line2D
+from matplotlib.patches import Patch
+
 try:
     import jax
     import jax.numpy as jnp
@@ -37,8 +40,224 @@ except ImportError:
     tfp_available = False
 
 
-__all__ = ["get_kernel", "get_mean", "get_prior", "get_log_likelihood", "GPResult", "get_gp_params"]
+__all__ = ["get_kernel", "get_mean", "get_prior", "get_log_likelihood", "GPResult", "get_gp_params","run_prior_checks"]
 
+
+def get_priors_samples(key, kernel_params, priors, loglike, n_samples=3000):
+    """Sample from the prior distribution.
+
+    Parameters
+    ----------
+    key : jax.random.PRNGKey
+        Random key.
+    priors : list
+        List of priors.
+    loglike : callable
+        Log likelihood function.
+    n_samples : int
+        Number of samples. Default is 3000.
+    """
+
+    num_params = len(kernel_params)
+
+    # get the prior model
+    prior_dict = dict(zip(kernel_params, priors))
+    prior_model = get_prior(kernel_params, prior_dict)
+
+    # define the model
+    nsmodel = Model(prior_model=prior_model, log_likelihood=loglike)
+    nsmodel.sanity_check(key=jax.random.PRNGKey(0), S=100)
+
+    # get the samples
+    unit_samples = jax.random.uniform(key, (n_samples, num_params))
+    prior_samples = jax.vmap(nsmodel.transform)(unit_samples)
+
+    return prior_samples
+
+
+def get_psd_and_approx(
+    kernel_type, kernel_params, prior_samples, f0, fM, n_frequencies=1000
+):  # -> tuple[NDArray[Any], NDArray[Any]]:
+    """Get the PSD and the approximate PSD for a given set of parameters and samples.
+
+    Parameters
+    ----------
+    n_samples : int
+        Number of samples.
+    kernel_params : list[str]
+        List of kernel parameters.
+    prior_samples : NDArray
+        Prior samples.
+    f0 : float
+        Minimum frequency.
+    fM : float
+        Maximum frequency.
+    n_frequencies : int
+        Number of frequencies.
+    """
+    n_samples = prior_samples[kernel_params[0]].shape[0]
+    f = np.geomspace(f0, fM, n_frequencies)
+    psd_models = []
+    psd_approx = []
+    for k in range(n_samples):
+        param_dict = {}
+        for i, params in enumerate(kernel_params):
+            if params[0:4] == "log_":
+                param_dict[params[4:]] = jnp.exp(prior_samples[params][k])
+            else:
+                param_dict[params] = prior_samples[params][k]
+
+        psd_model, psd_SHO = get_psd_approx_samples(f, kernel_type, param_dict, f0, fM)
+        psd_models.append(psd_model)
+        psd_approx.append(psd_SHO)
+    psd_models = np.array(psd_models)
+    psd_approx = np.array(psd_approx)
+    return f, psd_models, psd_approx
+
+
+def run_prior_checks(
+    kernel_type,
+    kernel_params,
+    priors,
+    loglike,
+    f_min,
+    f_max,
+    path="./",
+    n_samples=3000,
+    n_frequencies=1000,
+    key=jax.random.PRNGKey(42),
+    S_low=20,
+    S_high=20,
+):
[…]
