### 2024-07-19 653a028
remove the log of constant mean if logtransform of the data
