+++ stingray/modeling/gpmodeling.py
@@ -40,7 +40,15 @@ except ImportError:
     tfp_available = False
 
 
-__all__ = ["get_kernel", "get_mean", "get_prior", "get_log_likelihood", "GPResult", "get_gp_params","run_prior_checks"]
+__all__ = [
+    "get_kernel",
+    "get_mean",
+    "get_prior",
+    "get_log_likelihood",
+    "GPResult",
+    "get_gp_params",
+    "run_prior_checks",
+]
 
 
 def get_priors_samples(key, kernel_params, priors, loglike, n_samples=3000):
@@ -76,7 +84,14 @@ def get_priors_samples(key, kernel_params, priors, loglike, n_samples=3000):
 
 
 def get_psd_and_approx(
-    kernel_type, kernel_params, prior_samples, f0, fM, n_frequencies=1000
+    kernel_type,
+    kernel_params,
+    prior_samples,
+    f0,
+    fM,
+    n_frequencies=1000,
+    n_approx_components=20,
+    approximate_with="SHO",
 ):  # -> tuple[NDArray[Any], NDArray[Any]]:
     """Get the PSD and the approximate PSD for a given set of parameters and samples.
 
@@ -94,6 +109,7 @@ def get_psd_and_approx(
         Maximum frequency.
     n_frequencies : int
         Number of frequencies.
+
     """
     n_samples = prior_samples[kernel_params[0]].shape[0]
     f = np.geomspace(f0, fM, n_frequencies)
@@ -107,7 +123,15 @@ def get_psd_and_approx(
             else:
                 param_dict[params] = prior_samples[params][k]
 
-        psd_model, psd_SHO = get_psd_approx_samples(f, kernel_type, param_dict, f0, fM)
+        psd_model, psd_SHO = get_psd_approx_samples(
+            f,
+            kernel_type,
+            param_dict,
+            f0,
+            fM,
+            n_approx_components=n_approx_components,
+            approximate_with=approximate_with,
+        )
         psd_models.append(psd_model)
         psd_approx.append(psd_SHO)
     psd_models = np.array(psd_models)
@@ -128,9 +152,11 @@ def run_prior_checks(
     key=jax.random.PRNGKey(42),
     S_low=20,
     S_high=20,
+    n_approx_components=20,
+    approximate_with="SHO",
 ):
     """Check the approximation of the power spectrum density. This function will plot various diagnostics on the residuals and the ratio of the PSD and the approximate PSD.
-    
+
     Parameters
     ----------
     kernel_type : str
@@ -158,15 +184,22 @@ def run_prior_checks(
     S_low : int
         Low frequency scaling factor. Default is 20.
     S_high : int
-        High frequency scaling factor. Default is 20.    
+        High frequency scaling factor. Default is 20.
     """
     if kernel_type not in ["PowL", "DoubPowL"]:
         raise ValueError("Only 'PowL' and 'DoubPowL' kernels need to be checked")
-    
+
     f0, fM = f_min / S_low, f_max * S_high
     prior_samples = get_priors_samples(key, kernel_params, priors, loglike, n_samples)
     f, psd_models, psd_approx = get_psd_and_approx(
-        kernel_type, kernel_params, prior_samples, f0, fM, n_frequencies=n_frequencies
+        kernel_type,
+        kernel_params,
+        prior_samples,
+        f0,
+        fM,
+        n_frequencies=n_frequencies,
+        n_approx_components=n_approx_components,
+        approximate_with=approximate_with,
     )
 
     residuals = psd_approx - psd_models
[…]
