### 2024-08-01 35519a1
add posterior ppc and change kernel def
