+++ stingray/modeling/gpmodeling.py
@@ -47,7 +47,7 @@ __all__ = [
     "get_log_likelihood",
     "GPResult",
     "get_gp_params",
-    "run_prior_checks",
+    "run_prior_checks","run_posterior_check"
 ]
 
 
@@ -74,7 +74,7 @@ def get_priors_samples(key, kernel_params, priors, loglike, n_samples=3000):
 
     # define the model
     nsmodel = Model(prior_model=prior_model, log_likelihood=loglike)
-    nsmodel.sanity_check(key=jax.random.PRNGKey(0), S=100)
+    # nsmodel.sanity_check(key=jax.random.PRNGKey(0), S=1)
 
     # get the samples
     unit_samples = jax.random.uniform(key, (n_samples, num_params))
@@ -92,6 +92,7 @@ def get_psd_and_approx(
     n_frequencies=1000,
     n_approx_components=20,
     approximate_with="SHO",
+    with_normalisation=False,
 ):  # -> tuple[NDArray[Any], NDArray[Any]]:
     """Get the PSD and the approximate PSD for a given set of parameters and samples.
 
@@ -132,11 +133,136 @@ def get_psd_and_approx(
             n_approx_components=n_approx_components,
             approximate_with=approximate_with,
         )
-        psd_models.append(psd_model)
-        psd_approx.append(psd_SHO)
+        if with_normalisation:
+            f_c, a = _get_coefficients_approximation(
+                kernel_type,
+                param_dict,
+                f0,
+                fM,
+                n_approx_components=n_approx_components,
+                approximate_with=approximate_with,
+            )
+            norm = np.sum(a * f_c)
+            psd_models.append(psd_model * param_dict["variance"] / norm)
+            psd_approx.append(psd_SHO * param_dict["variance"] / norm)
+        else:
+            psd_models.append(psd_model)
+            psd_approx.append(psd_SHO)
     psd_models = np.array(psd_models)
     psd_approx = np.array(psd_approx)
     return f, psd_models, psd_approx
+
+
+def plot_psd_ppc(f, psd_quantiles, psd_approx_quantiles, psd_noise_levels, f_min, f_max, path):
+    """Replot the PSD PPC plot.
+
+    Parameters
+    ----------
+    f : array
+        The frequency array.
+    psd_quantiles : array
+        The quantiles of the PSD model.
+    psd_approx_quantiles : array
+        The quantiles of the PSD approximation.
+    psd_noise_levels : array
+        The noise levels.
+    f_min : float
+        The minimum frequency.
+    f_max : float
+        The maximum frequency.
+    path : str
+        The path to save the figure.
+    """
+    approx_color = "C6"
+    psd_color = "C3"
+    noise_color = "C5"
+    window_color = "k"
+
+    fig, ax = plt.subplots(1, 1, figsize=(6, 4))
+
+    ax.loglog(f, psd_quantiles[:, 2], label="Median", color=psd_color)
+    ax.fill_between(f, psd_quantiles[:, 1], psd_quantiles[:, 3], color=psd_color, alpha=0.3)
+    ax.fill_between(f, psd_quantiles[:, 0], psd_quantiles[:, 4], color=psd_color, alpha=0.15)
+    ax.axhline(psd_noise_levels[0], color=noise_color, ls="-", label="Noise level")
+    ax.loglog(f, psd_approx_quantiles[:, 2], color=approx_color)
+    ax.fill_between(
+        f,
+        psd_approx_quantiles[:, 1],
+        psd_approx_quantiles[:, 3],
[…]
