+++ stingray/modeling/tests/test_gpmodeling.py
@@ -22,8 +22,16 @@ try:
 except ImportError:
     _HAS_TINYGP = False
 
-from stingray.modeling.gpmodeling import get_kernel, get_mean, get_gp_params, _psd_model
-from stingray.modeling.gpmodeling import get_prior, get_log_likelihood, GPResult
+from stingray.modeling.gpmodeling import (
+    get_kernel,
+    get_mean,
+    get_gp_params,
+    _psd_model,
+    get_psd_and_approx,
+    run_prior_checks,
+    _get_coefficients_approximation,
+)
+from stingray.modeling.gpmodeling import get_prior, get_log_likelihood, GPResult, get_priors_samples
 from stingray import Lightcurve
 
 try:
@@ -46,7 +54,7 @@ def clear_all_figs():
         plt.close(fig)
 
 
-@pytest.mark.xfail
+# @pytest.mark.xfail
 @pytest.mark.skipif(not _HAS_TINYGP, reason="tinygp not installed")
 class Testget_kernel(object):
     def setup_class(self):
@@ -88,14 +96,19 @@ class Testget_kernel(object):
         ).all()
 
     def test_get_kernel_powL(self):
-        
-        t = jnp.arange(0,100.05,0.05)
-        f_min, f_max = 1/(t[-1]-t[0])/20, 1/(2*(t[1]-t[0]))*20
+
+        t = jnp.arange(0, 100.05, 0.05)
+        f_min, f_max = 1 / (t[-1] - t[0]) / 20, 1 / (2 * (t[1] - t[0])) * 20
         n_approx_components = 20
-        alpha_1, f_1, alpha_2,variance = 0.3, 0.05, 2.9,0.2
-        
-        kernel_params = {"alpha_1": alpha_1, "log_f_bend": jnp.log(f_1), "alpha_2": alpha_2, "variance": variance}
-        
+        alpha_1, f_1, alpha_2, variance = 0.3, 0.05, 2.9, 0.2
+
+        kernel_params = {
+            "alpha_1": alpha_1,
+            "f_bend": f_1,
+            "alpha_2": alpha_2,
+            "variance": variance,
+        }
+
         spectral_points = jnp.geomspace(f_min, f_max, 20)
         spectral_matrix = 1 / (
             1 + jnp.power(jnp.atleast_2d(spectral_points).T / spectral_points, 4)
@@ -116,10 +129,18 @@ class Testget_kernel(object):
             kernel += amplitudes[j] * kernels.quasisep.SHO(
                 quality=1 / jnp.sqrt(2), omega=2 * jnp.pi * spectral_points[j]
             )
-            
-        kernel_PowL_test = get_kernel("PowL", kernel_params,t,n_approx_components=n_approx_components)
-        assert (kernel(self.x, jnp.array([0.0])) == kernel_PowL_test(self.x, jnp.array([0.0]))).all()
-    
+
+        kernel_PowL_test = get_kernel(
+            "PowL",
+            kernel_params,
+            f_min=1 / (t[-1] - t[0]),
+            f_max=0.5 / (t[1] - t[0]),
+            n_approx_components=n_approx_components,
+        )
+        assert (
+            kernel(self.x, jnp.array([0.0])) == kernel_PowL_test(self.x, jnp.array([0.0]))
+        ).all()
+
     def test_value_error(self):
         with pytest.raises(ValueError, match="Kernel type not implemented"):
             get_kernel("periodic", self.kernel_params)
@@ -269,14 +290,14 @@ class Testget_gp_params(object):
         ]
 
     def test_get_params_PowL(self):
-        assert get_gp_params("PowL", "const") == [
+        assert get_gp_params("PowL", "constant") == [
             "alpha_1",
             "log_f_bend",
             "alpha_2",
             "variance",
[…]
