### 2024-08-05 cd615c2
Functions to split GTIs according to criteria
