### 2024-08-05 a13094d
Accept file names instead of hdulist
