### 2024-08-05 3262e70
Stub FITSTimeseriesReader implementation
