+++ stingray/io.py
@@ -16,6 +16,13 @@ from astropy.io import fits as pf
 import stingray.utils as utils
 from stingray.loggingconfig import setup_logger
 
+try:
+    import fitsio
+
+    _FITSIO_INSTALLED = True
+except ImportError:
+    _FITSIO_INSTALLED = False
+
 from .utils import (
     assign_value_if_none,
     is_string,
@@ -23,7 +30,13 @@ from .utils import (
     is_sorted,
     make_dictionary_lowercase,
 )
-from .gti import get_gti_from_all_extensions, load_gtis
+from .gti import (
+    get_gti_from_all_extensions,
+    load_gtis,
+    get_total_gti_length,
+    split_gtis_by_exposure,
+    cross_two_gtis,
+)
 
 from .mission_support import (
     read_mission_info,
@@ -712,6 +725,372 @@ def load_events_and_gtis(
 class EventReadOutput:
     def __init__(self):
         pass
+
+
+class FITSTimeseriesReader(object):
+    main_array_attr = "time"
+
+    def __init__(
+        self, fname, output_object_kind=None, gti_file=None, gtistring=None, additional_columns=None
+    ):
+        self.fname = fname
+        self._meta_attrs = []
+        self.gtistring = gtistring
+        self.output_object_kind = output_object_kind
+        self.additional_columns = additional_columns
+        self._initialize_header(fname)
+
+        if additional_columns is None and self.detector_key != "NONE":
+            additional_columns = [self.detector_key]
+        elif self.detector_key != "NONE":
+            additional_columns.append(self.detector_key)
+        self.data_hdu = fits.open(self.fname)[self.hduname]
+        self.gti_file = gti_file
+
+    @property
+    def time(self):
+        return self[:].time
+
+    def meta_attrs(self):
+        return self._meta_attrs
+
+    def add_meta_attr(self, name, value):
+        if name not in self._meta_attrs:
+            self._meta_attrs.append(name)
+        setattr(self, name, value)
+
+    @property
+    def exposure(self):
+        """
+        Return the total exposure of the time series, i.e. the sum of the GTIs.
+
+        Returns
+        -------
+        total_exposure : float
+            The total exposure of the time series, in seconds.
+        """
+
+        return get_total_gti_length(self.gti)
+
+    def __getitem__(self, index):
+        new_ts = self.output_object_kind()
+
+        columns = [self.time_column]
+        for col in self.pi_column, self.energy_column:
+            if col is not None:
+                columns.append(col)
+
+        data = self.data_hdu.data[index]
+
+        data = self._mission_specific_processing(data, header=self.header, hduname=self.hduname)
+
+        for attr in self.meta_attrs():
+            local_value = copy.deepcopy(getattr(self, attr))
+            if attr in ["t_start", "t_stop", "gti"] and local_value is not None:
+                setattr(new_ts, attr, local_value + self.timezero)
+            else:
+                setattr(new_ts, attr, local_value)
+
+        # Set the times
+        setattr(
+            new_ts,
+            self.main_array_attr,
+            data[self.time_column][:] + self.timezero,
+        )
+        # Get conversion function PI->Energy
+        try:
+            func = get_rough_conversion_function(
+                self.mission,
+                instrument=self.instr,
+                epoch=self.t_start / 86400 + self.mjdref,
+            )
+        except:
+            func = lambda x: x
+
+        if self.energy_column in data.dtype.names:
+            new_ts.energy = data[self.energy_column]
+        elif self.pi_column in data.dtype.names:
+            new_ts.pi = data[self.pi_column]
+            new_ts.energy = func(new_ts.pi)
+
+        det_numbers = None
+        if self.detector_key is not None and self.detector_key in data.dtype.names:
[…]
