### 2024-08-05 3ace4c5
Allow for shorter exposure cuts than GTIs
