### 2024-08-05 bf714fe
Fix empty function, better
