### 2024-08-12 #201 @hamogu
:wave: Hi @taldcroft and @masonng-astro!
Thanks for agreeing to do a review of stingray for us. Generic instructions (pre-review survey, etc.) are above: https://github.com/pyOpenSci/software-submission/issues/201#issuecomment-2204465146

In addition, note that this package is also submitting to JOSS and JOSS will accept our review and not do a separate review. However, stingray already had a JOSS publication a few years ago. So, it would be very helpful to note in your review if you deem the new features added since then as new "substantial contribution" that merits a new JOSS publication. You can find the paper draft for JOSS in https://github.com/StingraySoftware/stingray/pull/829

Please reach out to me if you have any questions!
