### 2024-09-04 839f851
Add functions to stream from timeseries
