### 2024-09-04 1f8013d
Add tests to eventlist
