### 2024-09-04 b8e845e
Allow to only stream attributes
