### 2024-09-05 9f49ae0
Use new machinery in analyze_segments
