### 2024-09-05 637ecf7
Use new machinery in split_by_gti
