### 2024-09-06 0e0c1c8
Fix problems with light curve slicing
