### 2024-09-06 #201 @masonng-astro
- [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

        - It looks like there is a list in “CREDITS.rst” but is not part of the metadata. There are no explicit contact details attached either.

    - [x]  Package has a [README.md](http://readme.md/) file in the root directory.

            - This is not present, as far as I can tell
            - This is not present, as far as I can tell
    - [x]  Descriptive links to all vignettes. If the package is small, there may only be a need for one vignette which could be placed in the [README.md](http://readme.md/) file.

    ### Usability

            - Almost every single test passed - see the code snippet at the very bottom of this review.
        - Not sure if this is explicitly noted, but as per above, there is no badge for which Python version are supported

    ### Final approval (post-review)

    ---

    As a frequent user of Stingray, I have casually followed the development of Stingray over the years. It has evolved from its previous iteration and from past projects (Stingray, HENDRICS, etc.) into a full software suite of spectral timing tools applicable to event data. I have always found the documentation to be extensive and comprehensive, with pedagogical Jupyter notebook-style tutorials to get started. My comments are thus fairly minor. I also note the absence of several items from the checklist above, though please let me know if I missed something.

    On the JOSS publication, I believe the addition of new features, the integration of additional data formats, and increased performance of the codebase merits a new JOSS publication. I just have very minor comments below.

    - Consistency between using “modelling” and “modeling”, except for the reference to the subpackage
    - “analyse” was written in the text instead of “analyze” (I assume the authors are adopting US English spelling throughout)
    - The hyperlinks to “our benchmarks” and “Lightkurve” point to wrong locations
    - I believe the authors will be adding the references before final approval

    I have glanced through all of the notebooks, and I encountered an issue in the “Bexvar tutorial” notebook. This might be my set-up, but I get the error that

```
python
    ImportError: dlopen(/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so, 0x0002): tried: '/Users/masonng/Documents/MIT/Research/software/heasoft-6.33/aarch64-apple-darwin22.1.0/mlfriends.cpython-310-darwin.so' (no such file), '/Users/masonng/Documents/MIT/Research/software/heasoft-6.33/aarch64-apple-darwin22.1.0/lib/mlfriends.cpython-310-darwin.so' (no such file), '/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64')), '/System/Volumes/Preboot/Cryptexes/OS/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (no such file), '/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64'))
```

    which I wonder if it is because there is some incompatibility with the arm64 architecture. I did install UltraNest with “arch -arm64” but the file “[mlfriends.cpython-310-darwin.so](http://mlfriends.cpython-310-darwin.so/)” is still in the x86_64 architecture. Apparently UltraNest does not explicitly support arm64 architecture ([[PyPI project here](https://www.notion.so/To-Do-List-f3febf42d06c41df81525237c183eb65?pvs=21)](https://www.notion.so/To-Do-List-f3febf42d06c41df81525237c183eb65?pvs=21))?

    I appreciate the “dependencies” sub-section under “installation instructions.” I would just like to include “UltraNest” to the list. Have there been tests done with the Apple Silicon chip-equipped devices as well? I encountered the above error with UltraNest for example, with my MacBook Pro M1; or this could be a personal problem.

```
    >>> stingray.test()
    =========================================================================================== test session starts ============================================================================================
    platform darwin -- Python 3.10.12, pytest-7.1.2, pluggy-1.5.0

    Running tests with stingray version 2.1.
    Running tests in /opt/homebrew/anaconda3/lib/python3.10/site-packages/stingray.

    Date: 2024-09-05T15:05:00

    Platform: macOS-13.0.1-arm64-arm-64bit

    Executable: /opt/homebrew/anaconda3/bin/python

    Full Python Version:
    3.10.12 | packaged by conda-forge | (main, Jun 23 2023, 22:41:52) [Clang 15.0.7 ]

    encodings: sys: utf-8, locale: UTF-8, filesystem: utf-8
    byteorder: little
    float info: dig: 15, mant_dig: 15

    Package versions:
    Numpy: 1.23.5
    Scipy: 1.10.0
    Matplotlib: 3.7.0
    h5py: 3.7.0
    scikit-image: 0.19.3

    Using Astropy options: remote_data: none.

    rootdir: /opt/homebrew/anaconda3/lib/python3.10/site-packages/stingray
    plugins: anyio-3.5.0, remotedata-0.4.1, doctestplus-1.2.1, astropy-header-0.2.2, asdf-3.4.0
    collected 1650 items / 1 skipped

    ../deadtime/tests/test_fad.py ................                                                                                                                                                       [  0%]
    ../deadtime/tests/test_models.py ...............                                                                                                                                                     [  1%]
    ../mission_support/tests/test_mission_support.py ..                                                                                                                                                  [  2%]
    ../modeling/tests/test_parameterestimation.py ............s.........s...............................                                                                                                 [  5%]
    ../modeling/tests/test_posterior.py .....................................................                                                                                                            [  8%]
    ../modeling/tests/test_scripts.py .......                                                                                                                                                            [  8%]
    ../pulse/overlapandsave/test_ols.py ss...                                                                                                                                                            [  9%]
    ../pulse/tests/test_accelsearch.py ......                                                                                                                                                            [  9%]
    ../pulse/tests/test_modeling.py ........                                                                                                                                                             [ 10%]
    ../pulse/tests/test_pulse.py .sss...............................                                                                                                                                     [ 12%]
    ../pulse/tests/test_search.py ..............................                                                                                                                                         [ 14%]
    ../simulator/tests/test_base.py ..                                                                                                                                                                   [ 14%]
    ../simulator/tests/test_simulator.py ........................................................                                                                                                        [ 17%]
    ../simulator/tests/test_transfer.py ................                                                                                                                                                 [ 18%]
    test_base.py .............................................................................................................................                                                           [ 26%]
    test_bexvar.py sssss..s.s.                                                                                                                                                                           [ 26%]
    test_bispectrum.py ............................                                                                                                                                                      [ 28%]
    test_covariancespectrum.py ......................                                                                                                                                                    [ 29%]
    test_crosscorrelation.py ...........................                                                                                                                                                 [ 31%]
    test_crossspectrum.py .............................................................................................................................................................................. [ 41%]
    ...........................................                                                                                                                                                          [ 44%]
    test_events.py ..............................................................                                                                                                                        [ 48%]
    test_filters.py ..........                                                                                                                                                                           [ 48%]
    test_fourier.py .................................................................................................................................................................................... [ 59%]
    ........                                                                                                                                                                                             [ 60%]
    test_gti.py .........................................................                                                                                                                                [ 63%]
    test_io.py ....F..............                                                                                                                                                                       [ 64%]
    test_lightcurve.py .....s..............................................................................................ss........................................ss.......                           [ 74%]
    test_lombscargle.py ...............................................                                                                                                                                  [ 76%]
    test_multitaper.py ......................................                                                                                                                                            [ 79%]
    test_power_colors.py .............                                                                                                                                                                   [ 80%]
    test_powerspectrum.py .................................................................................................................................................................              [ 89%]
    test_sampledata.py ....                                                                                                                                                                              [ 90%]
    test_spectroscopy.py xX......                                                                                                                                                                        [ 90%]
    test_stats.py .......................................................                                                                                                                                [ 93%]
    test_utils.py ...................................................s..........                                                                                                                         [ 97%]
    test_varenergyspectrum.py ........................................                                                                                                                                   [100%]

    ================================================================================================= FAILURES =================================================================================================
    ______________________________________________________________________________ TestIO.test_event_file_read_and_automatic_sort ______________________________________________________________________________

    self = <stingray.tests.test_io.TestIO object at 0x167545360>

        def test_event_file_read_and_automatic_sort(self):
            """Test event file reading."""
            fname = os.path.join(datadir, "monol_testA_calib.evt")
            with pytest.warns(AstropyUserWarning, match="No valid GTI extensions"):
                evdata = load_events_and_gtis(fname)
            fname_unsrt = os.path.join(datadir, "monol_testA_calib_unsrt.evt")
    >       with pytest.warns(UserWarning, match="not sorted. Sorting them for you"):
    E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted. The list of emitted warnings is: [].

    test_io.py:77: Failed
    ============================================================================================= warnings summary =============================================================================================
    ../../_pytest/config/__init__.py:1129
      /opt/homebrew/anaconda3/lib/python3.10/site-packages/_pytest/config/__init__.py:1129: PytestAssertRewriteWarning: Module already imported so cannot be rewritten: asdf
        self._mark_plugins_for_rewrite(hook)

    ../deadtime/tests/test_fad.py:22: 1 warning
    ../deadtime/tests/test_models.py:14: 1 warning
    ../modeling/tests/test_parameterestimation.py:34: 1 warning
    ../pulse/overlapandsave/test_ols.py:88: 1 warning
    ../pulse/tests/test_accelsearch.py:6: 1 warning
    ../pulse/tests/test_pulse.py:32: 1 warning
    ../pulse/tests/test_pulse.py:55: 1 warning
    ../pulse/tests/test_search.py:14: 1 warning
    test_bexvar.py:11: 1 warning
    test_crossspectrum.py:331: 1 warning
    test_crossspectrum.py:909: 1 warning
    test_crossspectrum.py:921: 1 warning
    test_crossspectrum.py:938: 1 warning
    test_crossspectrum.py:958: 1 warning
    test_lightcurve.py:1605: 1 warning
    test_multitaper.py:13: 1 warning
    test_powerspectrum.py:210: 1 warning
    test_spectroscopy.py:93: 1 warning
    test_varenergyspectrum.py:171: 1 warning
    test_varenergyspectrum.py:358: 1 warning
      PytestUnknownMarkWarning: Unknown pytest.mark.slow - is this a typo?  You can register custom marks to avoid this warning - for details, see <https://docs.pytest.org/en/stable/how-to/mark.html>

    ../../pint/pulsar_mjd.py:57
      RuntimeWarning: This platform does not support extended precision floating-point, and PINT will run at reduced precision.

    deadtime/tests/test_fad.py: 2 warnings
    tests/test_base.py: 3 warnings
    tests/test_crossspectrum.py: 1 warning
    tests/test_events.py: 2 warnings
    tests/test_lightcurve.py: 1 warning
    tests/test_powerspectrum.py: 1 warning
    tests/test_varenergyspectrum.py: 1 warning
      UserWarning: table path was not set via the path= argument; using default path __astropy_table__

    deadtime/tests/test_models.py::test_checkA[True]
    deadtime/tests/test_models.py::test_checkA[False]
      UserWarning: All values for SymLogScale are below linthresh, making it effectively linear. You likely should lower the value of linthresh.

    modeling/tests/test_parameterestimation.py::TestParameterEstimation::test_sampler_runs
      DeprecationWarning: The py23 module has been deprecated and will be removed in a future release. Please update your code.

    modeling/tests/test_parameterestimation.py::TestPSDParEst::test_fit_method_works_with_correct_parameter
      RuntimeWarning: More than 20 figures have been opened. Figures created through the pyplot interface (`matplotlib.pyplot.figure`) are retained until explicitly closed and may consume too much memory. (To control this warning, see the rcParam `figure.max_open_warning`). Consider using `matplotlib.pyplot.close()`.

    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_crossspectrum
    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_powerspectrum
    pulse/tests/test_accelsearch.py::TestAccelsearch::test_signal
    pulse/tests/test_accelsearch.py::TestAccelsearch::test_signal_neg_fdot
    pulse/tests/test_accelsearch.py::TestAccelsearch::test_noisy
    pulse/tests/test_accelsearch.py::TestAccelsearch::test_noisy_neg_fdot
      RuntimeWarning: divide by zero encountered in log

    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_crossspectrum
    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_powerspectrum
      RuntimeWarning: divide by zero encountered in divide

    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_crossspectrum
    modeling/tests/test_scripts.py::TestFitLorentzians::test_fit_powerspectrum
      RuntimeWarning: invalid value encountered in subtract

    pulse/tests/test_modeling.py::test_gaussian_bounds
      RuntimeWarning: underflow encountered in exp

    pulse/tests/test_modeling.py::test_gaussian_bounds
      RuntimeWarning: underflow encountered in multiply

    pulse/tests/test_search.py: 1 warning
    simulator/tests/test_simulator.py: 39 warnings
    tests/test_crosscorrelation.py: 1 warning
    tests/test_crossspectrum.py: 2 warnings
    tests/test_lightcurve.py: 1 warning
    tests/test_lombscargle.py: 3 warnings
    tests/test_multitaper.py: 1 warning
    tests/test_varenergyspectrum.py: 3 warnings
      UserWarning: SIMON says: Stingray only uses poisson err_dist at the moment. All analysis in the light curve will assume Poisson errors. Sorry for the inconvenience.

    simulator/tests/test_simulator.py: 2 warnings
    tests/test_crosscorrelation.py: 1 warning
    tests/test_crossspectrum.py: 75 warnings
    tests/test_spectroscopy.py: 32 warnings
    tests/test_varenergyspectrum.py: 1 warning
      UserWarning: n_ave is below 30. Please note that the error bars on the quantities derived from the cross spectrum are only reliable for a large number of averaged powers.

    simulator/tests/test_simulator.py: 2 warnings
    tests/test_crosscorrelation.py: 1 warning
    tests/test_crossspectrum.py: 97 warnings
    tests/test_lombscargle.py: 10 warnings
    tests/test_spectroscopy.py: 46 warnings
    tests/test_varenergyspectrum.py: 2 warnings
      RuntimeWarning: invalid value encountered in sqrt

    tests/test_bexvar.py::TestInternalFunctions::test_lscg_gen
[…]
