+++ docs/changes/838.feature.rst
@@ -1,0 +1,1 @@
+Add methods to stream timeseries data into chunks
