### 2024-09-09 ee97cc8
Better test of all cases when getting gtis from all extensions
