### 2024-09-09 StingraySoftware/stingray#841 @matteobachetti
From pyOpenSci review: better deal with UltraNest dependency
### Description of the Issue

_Originally posted by @masonng-astro in https://github.com/pyOpenSci/software-submission/issues/201#issuecomment-2334902673_:

I have glanced through all of the notebooks, and I encountered an issue in the “Bexvar tutorial” notebook. This might be my set-up, but I get the error that

    ```python
    ImportError: dlopen(/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so, 0x0002): tried: '/Users/masonng/Documents/MIT/Research/software/heasoft-6.33/aarch64-apple-darwin22.1.0/mlfriends.cpython-310-darwin.so' (no such file), '/Users/masonng/Documents/MIT/Research/software/heasoft-6.33/aarch64-apple-darwin22.1.0/lib/mlfriends.cpython-310-darwin.so' (no such file), '/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64')), '/System/Volumes/Preboot/Cryptexes/OS/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (no such file), '/opt/homebrew/anaconda3/lib/python3.10/site-packages/ultranest/mlfriends.cpython-310-darwin.so' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64'))
    ```

which I wonder if it is because there is some incompatibility with the arm64 architecture. I did install UltraNest with “arch -arm64” but the file “mlfriends.cpython-310-darwin.so” is still in the x86_64 architecture. Apparently UltraNest does not explicitly support arm64 architecture ([PyPI project here](https://www.notion.so/To-Do-List-f3febf42d06c41df81525237c183eb65?pvs=21))?

I appreciate the “dependencies” sub-section under “installation instructions.” I would just like to include “UltraNest” to the list. Have there been tests done with the Apple Silicon chip-equipped devices as well? I encountered the above error with UltraNest for example, with my MacBook Pro M1; or this could be a personal problem.
