### 2024-09-09 #201 @matteobachetti
@cmarmo @lwasser in the README tutorial at https://www.pyopensci.org/python-package-guide/tutorials/add-readme.html I can't find a recommended way to specify the supported Python versions. Is simple textual information enough?
