### 2024-09-09 #201 @taldcroft
- This information is in `setup.cfg`

      - Not included. I somewhat question the utility of this badge since it seems to rely
        on the package maintainers to keep it up to date. An abandoned project is not
        likely to be marked as such.
      - Looks like Py 3.8 to 3.12 are tested, but I do not see this in the install docs.

  - I do not have the domain knowledge or resources to confirm the functional claims of
    the software with regards to X-ray spectral timing. The extensive collection of
    notebooks should allow the motivated scientist to evaluate and understand the
    results from this package.
  - Documentation makes no specific performance claims.

---

I am reviewing this package as an (inactive) astronomer and active Astropy developer.

From a scientific perspective it appears that the `stingray` package fills a distinct
need for modern open source Python-based X-ray spectral timing tools. It is being widely
used in the X-ray timing community, with 170 citations to the 2019 ApJ paper. In recent
years it is seeing around 35 citations per year.

From a software perspective this checks all the boxes (literally) for a package that
meets current standards for a well-developed, documented, and maintained package. The
test coverage is excellent.

I ran all the notebooks and found a couple of problems (documented as issues). This
highlights the need for https://github.com/StingraySoftware/stingray/issues/802 to be
worked, but overall things are looking OK and those notebook issues are not a blocker
for accepting the package.

I opened one issue to document how to download the tutorial notebooks that form most of
the package functional documentation. **This is a simple docs update and I would like to
see this done for accepting the package.**
