### 2024-09-10 #201 @taldcroft
FWIW as noted in my review, tests passed for me on Mac Silicon and Python 3.12. I created a fresh conda environment (using conda defaults channel) for Python 3.12 and then pip installed everything else.
