### 2024-09-28 #201 @hamogu
@masonng-astro Could you have a look if your concerns have been resolved or if you can provide a more detailed look at why it fails? It seems that @matteobachetti tried to reproduce your failure, but failed to do so. Of course, we don't want to accept a failing package, but it's also hard to figure out what went wrong. 😕

Once you think it's ready to be accepted, please edit the post with your review and tick the box on "I recommend approving this package." and leave. short message here (GH will inform me when you post a new comment, but not when you edit an existing one).
