### 2024-10-01 #201 @masonng-astro
> @masonng-astro I have a similar setup to yours (Macbook pro with M1 processor), but I couldn't reproduce your bugs 🆘
>
> The failing test on unsorted data is particularly weird, as it does not seem to rely on any specific dependency other than the default ones. There is a vague chance that it is related to the numba installation, as I use numba to check for sortedness, but it should also revert to standard numpy if numba is not installed or it fails.
>
> I could also install UltraNest without issues. I wonder if you have conflicting versions of some dependencies? From the list of libraries in your error, there seems to be some kind of clash between HEASOFT-installed libraries and the standard Python path. Is it possible?

Sorry for my late response - I did eventually manage to hunt down the issue and it boiled down to a weird esoteric mix of architectures in my Anaconda installation which I never encountered until now! Apologies for holding up the review. I'll go ahead and approve now. I did check everything else and all is good on my end!
