### 2024-10-01 #201 @matteobachetti
@hamogu I opened a PR with the badge, and filled in the form. Later on I will open the JOSS issue. Thanks a lot!
