### 2024-10-01 #201 @matteobachetti
> @matteobachetti We at pyOpenSci would love to highlight stingray in a blog post similar to what these packages have done: [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation), [movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis) ([We have an md template](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) that you could use as a guide when creating your post.) but I know that you've written a lot about Stingray before, and this is certainly not a requirement!

I think it's a good idea, I have some matters to deal with in the next few weeks but I'll try to do it. Thanks!
>
> I also want to invite all three of you to the PyOpenSci slack - there is a very active community that discusses best practices in packaging, testing , and administration of Python packages. Personally, I found a lot of those discussions very insightful and useful, thought I myself have never contributed by posting - I just use it as a resource to look at. If you'd like to join, let me know which email address I should invite!

Sent in private message!
