+++ _sources/_zenodo.rst.txt
@@ -1,0 +1,48 @@
+.. list-table::
+   :header-rows: 1
+
+   * - Stingray Release
+     - DOI
+     - Citation
+   * - `v2.2 <https://github.com/StingraySoftware/stingray/releases/tag/v2.2>`__
+     - `10.5281/zenodo.13974481 <https://zenodo.org/records/13974481>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/13974481/export/hx>`__
+   * - `v2.1 <https://github.com/StingraySoftware/stingray/releases/tag/v2.1>`__
+     - `10.5281/zenodo.11383212 <https://zenodo.org/records/11383212>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/11383212/export/hx>`__
+   * - `v2.0.0rc1 <https://github.com/StingraySoftware/stingray/releases/tag/v2.0.0rc1>`__
+     - `10.5281/zenodo.10604413 <https://zenodo.org/records/10604413>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/10604413/export/hx>`__
+   * - `v2.0.0 <https://github.com/StingraySoftware/stingray/releases/tag/v2.0.0>`__
+     - `10.5281/zenodo.10813181 <https://zenodo.org/records/10813181>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/10813181/export/hx>`__
+   * - `v1.1.2 <https://github.com/StingraySoftware/stingray/releases/tag/v1.1.2>`__
+     - `10.5281/zenodo.7970570 <https://zenodo.org/records/7970570>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/7970570/export/hx>`__
+   * - `v1.1 <https://github.com/StingraySoftware/stingray/releases/tag/v1.1>`__
+     - `10.5281/zenodo.7135161 <https://zenodo.org/records/7135161>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/7135161/export/hx>`__
+   * - `v1.0-beta <https://github.com/StingraySoftware/stingray/releases/tag/v1.0-beta>`__
+     - `10.5281/zenodo.6290078 <https://zenodo.org/records/6290078>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/6290078/export/hx>`__
+   * - `v1.0 <https://github.com/StingraySoftware/stingray/releases/tag/v1.0>`__
+     - `10.5281/zenodo.6394742 <https://zenodo.org/records/6394742>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/6394742/export/hx>`__
+   * - `v0.3 <https://github.com/StingraySoftware/stingray/releases/tag/v0.3>`__
+     - `10.5281/zenodo.4881255 <https://zenodo.org/records/4881255>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/4881255/export/hx>`__
+   * - `v0.2 <https://github.com/StingraySoftware/stingray/releases/tag/v0.2>`__
+     - `10.5281/zenodo.3898435 <https://zenodo.org/records/3898435>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/3898435/export/hx>`__
+   * - `v0.1.3 <https://github.com/StingraySoftware/stingray/releases/tag/v0.1.3>`__
+     - `10.5281/zenodo.3242835 <https://zenodo.org/records/3242835>`__
+     - `[Link to BibTeX] <https://zenodo.org/records/3242835/export/hx>`__
+   * - `v0.1.2 <https://github.com/StingraySoftware/stingray/releases/tag/v0.1.2>`__
[…]
