### 2024-10-24 402b536
Textbook-like implementation of sampling. In this implementation, phase is taken into account. The intensity is 0.5 at the edges.
