### 2024-11-09 80604cf
Modify apply_gti_lists method in FITSTimeseriesReader class
