### 2024-11-13 #201 @hamogu
I see that there is a JOSS DOI now, so I've updated all the information and will close this issue. Thanks everyone!
